scrum/requirements.txt
djangorestframework==3.1.3
Markdown==2.6.2
django-filter==0.10.0
psycopg2==2.6.1